Merge dev

+++ src/objc/tests/common/random/Random.mm
@@ -314,7 +314,7 @@
     "0123456789"
     "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
     "abcdefghijklmnopqrstuvwxyz";
-    char result[std::numeric_limits<uint8_t>::max() + 1];
+    char result[std::numeric_limits<uint16_t>::max() + 1];
     for (int i = 0; i < length; ++i) {
         result[i] = alphanum[self.uint8 % (sizeof(alphanum) - 1)];
     }
